# Conflicts:
#	src/main/java/hhplus/concert/domain/pay/PaymentService.java
#	src/main/java/hhplus/concert/infra/persistence/ConcertRepositoryImpl.java
#	src/main/java/hhplus/concert/infra/persistence/ReservationJpaRepository.java
#	src/main/java/hhplus/concert/support/config/RedisConfig.java

+++ src/main/java/hhplus/concert/domain/userqueue/UserQueueConstant.java
@@ -6,10 +6,10 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @Getter
 @ConfigurationProperties(prefix = "user-queue")
 public class UserQueueConstant {
-    private final Integer maxWaitingNumber;
-    private final Integer queueTokenExpireTime;
+    private final Long maxWaitingNumber;
+    private final Long queueTokenExpireTime;
 
-    public UserQueueConstant(Integer maxWaitingNumber, Integer queueTokenExpireTime) {
+    public UserQueueConstant(Long maxWaitingNumber, Long queueTokenExpireTime) {
         this.maxWaitingNumber = maxWaitingNumber;
         this.queueTokenExpireTime = queueTokenExpireTime;
     }
